requirements.txt
builtwith==1.3.4
dnspython==2.8.0
ipwhois==1.3.0
PythonDNS==0.1
requests==2.32.5